fix(M-15,L-10): correct bbox quaternion order/value; solve PnP in float64
M-15: bbox.json5 commented the rotation as (w, x, y, z) but the node parses
nalgebra's (x, y, z, w) order. Under (x,y,z,w) the shipped value [1,0,0,0] is a
180-deg rotation about X, not the identity -- masked only because the box is
symmetric in y and z. Corrected the comment and set the value to the true identity
[0,0,0,1]; the filtered volume is unchanged. Anyone hand-writing a *rotated* bbox
(as the field captures do) was being pointed at the wrong volume, then getting
silent empty detections (C-04).

L-10: the advanced solver cast every PnP input to float32 while the rest of the
pipeline is float64. Use float64 throughout the solve -- OpenCV accepts CV_64F and
there is no performance argument at 16*N points -- so the isolated precision
downgrade is gone and a future cond(JtJ) diagnostic stays trustworthy.

Verified: bbox parses to identity [0,0,0,1]; no float32 remains in the solver;
just build (2+13, 0 fail), just test (273 Rust, 48 Python).

Refs: docs/issues/M-15-bbox-quaternion-order-comment.md,
      docs/issues/L-10-solver-float32-precision.md

+++ ros/advanced_extrinsic_solver/advanced_extrinsic_solver/main.py
@@ -1088,8 +1088,8 @@ class AdvancedExtrinsicSolver(Node):
             all_image_points.extend(image_points)
 
         # Convert to numpy arrays
-        all_object_points = np.array(all_object_points, dtype=np.float32)
-        all_image_points = np.array(all_image_points, dtype=np.float32)
+        all_object_points = np.array(all_object_points, dtype=np.float64)
+        all_image_points = np.array(all_image_points, dtype=np.float64)
 
         num_correspondences = len(all_object_points)
 
@@ -1391,9 +1391,9 @@ class AdvancedExtrinsicSolver(Node):
         image_points = []
 
         board_rotation = (
-            R.from_quat(board_detection.orientation).as_matrix().astype(np.float32)
+            R.from_quat(board_detection.orientation).as_matrix().astype(np.float64)
         )
-        board_position = np.array(board_detection.position, dtype=np.float32)
+        board_position = np.array(board_detection.position, dtype=np.float64)
 
         board_frame_corners = self._compute_multi_marker_corners()
 
@@ -1414,16 +1414,16 @@ class AdvancedExtrinsicSolver(Node):
                 )
                 continue
 
-            local_corners = np.array(board_frame_corners[marker_id], dtype=np.float32)
+            local_corners = np.array(board_frame_corners[marker_id], dtype=np.float64)
             world_corners = (board_rotation @ local_corners.T).T + board_position
 
-            image_corners = np.array(marker.corners, dtype=np.float32)
+            image_corners = np.array(marker.corners, dtype=np.float64)
 
             object_points.extend(world_corners)
             image_points.extend(image_corners)
 
-        return np.array(object_points, dtype=np.float32), np.array(
-            image_points, dtype=np.float32
+        return np.array(object_points, dtype=np.float64), np.array(
+            image_points, dtype=np.float64
         )
 
     def _solve_pnp(
@@ -1434,8 +1434,8 @@ class AdvancedExtrinsicSolver(Node):
             self.get_logger().error("PnP requires at least 4 point correspondences")
             return False, None, None
 
-        K = np.array(self.camera_info.k, dtype=np.float32).reshape(3, 3)
-        dist_coeffs = np.zeros(5, dtype=np.float32)
+        K = np.array(self.camera_info.k, dtype=np.float64).reshape(3, 3)
+        dist_coeffs = np.zeros(5, dtype=np.float64)
 
         self.get_logger().info(f"Solving PnP with {len(object_points)} correspondences")
